fix: forward shell env to pager
This patch fixes the motd.pager mode to start launching
the pager with the same environment that we launch the
shell with. The old way of just inheriting the same
environment that the daemon has has the problem that
it will not correctly match the user's TERM value, which
can cause problems for programs like `less` that use
the TERM value to determine the right control sequences
to emit.

Fixes #129